requirement.txt
altgraph==0.17
aniso8601==9.0.1
asgiref==3.5.2
attrs==21.2.0
auto-py-to-exe==2.7.11
autopep8==1.6.0
beautifulsoup4==4.10.0
bottle==0.12.19
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.11.0
Django==4.1
django-cors-headers==3.13.0
django-filter==23.2
djangorestframework==3.14.0
dtw==1.4.0
easyocr==1.6.2
Eel==0.12.4
filelock==3.10.0
Flask==2.0.1
flask-restx==0.5.0
Flask-WTF==0.15.1
fonttools==4.33.3
future==0.18.2
gensim==4.0.1
gevent==20.9.0
gevent-websocket==0.10.1
googletrans==4.0.0rc1
greenlet==0.4.17
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.26.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
JPype1 @ file:///D:/Downloads/JPype1-1.4.0-cp39-cp39-win_amd64.whl
jsonschema==3.2.0
kiwisolver==1.4.2
lazy_loader==0.1
MarkupSafe==2.0.1
matplotlib==3.5.2
MouseInfo==0.1.3
mpmath==1.3.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
networkx==3.0
ninja==1.11.1
numpy==1.24.2
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.3.2
pefile==2019.4.18
Pillow==9.0.1
PyAutoGUI==0.9.53
pyclipper==1.3.0.post4
pycodestyle==2.8.0
pycparser==2.20
pycryptodomex==3.10.4
PyGetWindow==0.0.9
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==3.0.8
pypdf==3.1.0
PyPDF2==3.0.0
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.0
PyScreeze==0.1.28
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.1
PyWavelets==1.4.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.26.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
scikit-image==0.20.0
scikit-learn==0.24.2
scipy==1.9.1
shapely==2.0.1
six==1.16.0
smart-open==5.2.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
sympy==1.11.1
threadpoolctl==2.2.0
tifffile==2023.3.15
toml==0.10.2
torch==2.0.0
torchvision==0.15.1
typing_extensions==4.4.0
tzdata==2021.5
urllib3==1.26.6
Werkzeug==2.0.1
whichcraft==0.6.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.2.0
